pyproject.toml
[project]
name = "flowledger"
version = "0.1.0"
description = ""
authors = [
    {name = "Tzur Vaich",email = "tzurvaich@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.135.3,<0.136.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.fastapi]
entrypoint = "app.main:app"
